[ASV-1211] Moved autoupdate files to the respective package
