+++ app/src/main/java/cm/aptoide/pt/autoupdate/AutoUpdateJsonResponse.java
@@ -1,4 +1,4 @@
-package cm.aptoide.pt.view;
+package cm.aptoide.pt.autoupdate;
 
 public class AutoUpdateJsonResponse {
